relocate postRecieve to internal/metadata
